Hold psyq-asm to a real-assembler corpus
25 original C programs in test/fixtures/corpus/src cover constructs the
psyq-wasm fixtures do not: dense and sparse switches, long long, floats,
struct copies, bitfields, varargs, big frames, hardware addresses, globals
of every size, statics, computed goto, and more. scripts/compile-corpus.mjs
compiles them at -G 0, -G 8, and -G 8 -g (CI checks the output with
npm run corpus:verify), and every one of the 75 files carries the words,
section sizes, relocations, and symbols the real ASPSX 2.81 emits, which the
tests require.

What the corpus found, now fixed:
- GNU numeric local labels (1:, 1f, 1b), emitted in long long code, were
  unsupported; they now resolve and are transparent to the nop pass.
- .lcomm aligns to its size rounded up to a power of two, at most 16, not
  capped at 8 (new probe VERIFY-23 sweeps 24 sizes).
- The object reader read SECTION_DEF2 dimensions as two bytes; they are
  four.
- With -g, ASPSX also writes local symbols for static data, so recorded
  locals are compared as required and permitted sets.

Next: a seeded generator to fuzz psyq-asm against ASPSX.

+++ test/unit/psyq-object.test.ts
@@ -60,7 +60,7 @@ const SYNTHETIC = object(
   [78, new Array<number>(10).fill(0)],
   [80, new Array<number>(10).fill(0)],
   [82, new Array<number>(14).fill(0), 'def'],
-  [84, new Array<number>(14).fill(0), u16(2), u16(3), u16(4), 'tag', 'def2'],
+  [84, new Array<number>(14).fill(0), u16(2), u32(3), u32(4), 'tag', 'def2'],
   [86, new Array<number>(36).fill(0), 'probe2'],
   [0],
   [0xff], // after END: ignored
